requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
requests==2.31.0
